// server/src/gameplay/gameOne/interfaces/PlayerStateForClient.ts
import { IPlayerState } from '../../interfaces';
import { Obstacle } from './';

export interface PlayerStateForClient extends IPlayerState {
    id: string;
    name: string;
    positionX: number;
    obstacles: Array<Obstacle>;
    atObstacle: boolean;
    finished: boolean;
    finishedTimeMs: number;
    dead: boolean;
    rank: number;
    isActive: boolean;
    stunned: boolean;
    characterNumber: number;
    chaserPushesUsed: number;
}